refactor: changed evaluationWeightUtils.ts to evaluationWeightTracker.ts and move to EvaluationWeightTracker folder under Evaluation folder
Moved the weight tracking utility to the Evaluation module for better cohesion and structure.
Updated import paths will be handled in corresponding modules as needed.
